requirements.txt
fastapi
uvicorn
pydantic
langgraph
langchain-groq
langchain-core
python-dotenv
duckduckgo-search
langchain-community
streamlit
requests